Prepare for CI.
