--- GamepadCalibrator/MakefileLinux
@@ -1,10 +1,0 @@
-# Constants
-SIMPLEINI_DIR = "/home/mike/repo/simpleini"
-
-INCLUDE_DIR_FLAGS = -I${SIMPLEINI_DIR}
-
-FLAGS_MUTUAL = -Wall -std=c++11  ${INCLUDE_DIR_FLAGS}
-
-all: 
-	g++ ${FLAGS_MUTUAL} GamepadCalibrator.cpp joystick.cc -o Calibrator
-
