+++ calibrator/config.ini
@@ -1,0 +1,104 @@
+[axis]
+Exit = 9
+B1 = 1
+B2 = 2
+B3 = 3
+B4 = 4
+L1 = 7
+L2 = 5
+R1 = 8
+R2 = 6
+start = 12
+T1 = 10
+T2 = 11
+RTUD = 16
+RTLR = 15
+LTUD = 14
+LTLR = 13
+arrowsUD = 14
+arrowsLR = 17
+
+
+[B1]
+upper_value = 1
+lower_value = 0
+
+
+[B2]
+upper_value = 1
+lower_value = 0
+
+
+[B3]
+upper_value = 1
+lower_value = 0
+
+
+[B4]
+upper_value = 1
+lower_value = 0
+
+
+[L1]
+upper_value = 1
+lower_value = 0
+
+
+[L2]
+upper_value = 1
+lower_value = 0
+
+
+[R1]
+upper_value = 1
+lower_value = 0
+
+
+[R2]
+upper_value = 1
+lower_value = 0
+
+
+[start]
+upper_value = 1
+lower_value = 0
+
+
+[T1]
+upper_value = 1
+lower_value = 0
+
+
+[T2]
+upper_value = 1
+lower_value = 0
+
+
+[RTUD]
+upper_value = -32767
+lower_value = 32767
+
+
+[RTLR]
+upper_value = -32767
+lower_value = 32767
+
+
+[LTUD]
+upper_value = -32767
+lower_value = 32767
+
+
+[LTLR]
+upper_value = -32767
+lower_value = -32767
+
+
+[arrowsUD]
+upper_value = -23985
+lower_value = 32767
+
+
+[arrowsLR]
+upper_value = -32767
+lower_value = 32767
