+++ module/config.ini
@@ -1,0 +1,108 @@
+[options]
+input_path = /dev/input/js1
+
+[axis]
+Exit = 9
+B1 = 1
+B2 = 2
+B3 = 3
+B4 = 4
+L1 = 7
+L2 = 5
+R1 = 8
+R2 = 6
+start = 10
+T1 = 11
+T2 = 12
+RTUD = 13
+RTLR = 16
+LTUD = 15
+LTLR = 14
+arrowsUD = 17
+arrowsLR = 18
+
+
+[B1]
+upper_value = 1
+lower_value = 0
+
+
+[B2]
+upper_value = 1
+lower_value = 0
+
+
+[B3]
+upper_value = 1
+lower_value = 0
+
+
+[B4]
+upper_value = 1
+lower_value = 0
+
+
+[L1]
+upper_value = 1
+lower_value = 0
+
+
+[L2]
+upper_value = 1
+lower_value = 0
+
+
+[R1]
+upper_value = 1
+lower_value = 0
+
+
+[R2]
+upper_value = 1
+lower_value = 0
+
+
+[start]
+upper_value = 1
+lower_value = 0
+
+
+[T1]
+upper_value = 1
+lower_value = 0
+
+
+[T2]
+upper_value = 1
+lower_value = 0
+
+
+[RTUD]
+upper_value = 0
+lower_value = 65535
+
+
+[RTLR]
+upper_value = 0
+lower_value = 65535
+
+
+[LTUD]
+upper_value = 0
+lower_value = 65535
+
+
+[LTLR]
+upper_value = 0
+lower_value = 65535
+
+
+[arrowsUD]
+upper_value = 1
+lower_value = -1
+
+
+[arrowsLR]
+upper_value = 1
+lower_value = -1
+
